unfixed-bug: could not get the python process to run
Trying to have the script run as soon as the server is started
however, the data is never processed.  There are two potential
problems I think this might be.

(1) async issues: the write function defined on line 21, is executed
but is never waited on.

(2) a python compilation issue that I am unaware of.